--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -42,6 +42,7 @@
 - PR #1846 C++ type-erased gdf_equal_columns test util; fix gdf_equal_columns logic error
 - PR #1882 Add iloc functionality to MultiIndex dataframes
 - PR #1896 Improve performance of groupby with levels specified in dask-cudf
+- PR #TODO Improve iloc performance for non-contiguous row selection
 
 ## Bug Fixes