This PR brings the following changes/improvements:

* Uses multi-column `gather` from libcudf rather than a gather on each column of the DataFrame (adds support for NVCategory in `cudf::gather`)
* Improves performance of iloc by operating on a view of the dataframe rather than a copy when selecting columns
* Makes adding columns to a DataFrame faster by avoiding the expensive check for equality of index

Before:

```python
In [4]: nrows = 10_000
In [5]: ncols = 1000
In [6]: data = np.random.rand(nrows, ncols)
In [7]: pdf = pd.DataFrame(data)
In [8]: gdf = cudf.from_pandas(pdf)
In [10]: rows = [x for x in range(nrows) if x%2 == 0]
In [11]: %timeit gdf.iloc[rows, :]
6.3 s ± 32.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

After:

```python
In [5]: nrows = 10_000
In [6]: ncols = 1000
In [7]: data = np.random.rand(nrows, ncols)
In [8]: pdf = pd.DataFrame(data)
In [9]: gdf = cudf.from_pandas(pdf)
In [11]: rows = [x for x in range(nrows) if x%2 == 0]
In [12]: %timeit gdf.iloc[rows, :]
906 ms ± 8.88 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

Closes #1880
Closes #1732