@kkraus14 2019-06-08
Python side is good to go, will let @harrism review again before merging